Add user-based approach to each service API key. Update Backend and Frontend. Move Schema into separate file. Add initialization endpoints for unintialized database.

+++ app.py
@@ -3,16 +3,26 @@ import uvicorn
 from fastapi import FastAPI, HTTPException, Depends, UploadFile, File, Body
 from fastapi.security import OAuth2PasswordRequestForm
 
-from configs.database_config import (DatabaseConfig, User, UserUpdate,
-                                     Completions, ChatCompletions,
-                                     Embeddings, FineTunes)
+from configs.database_config import DatabaseConfig
 from configs.service_config import ServiceConfig
 from configs.openai_config import OpenAIConfig
+from configs.authentication_config import AuthenticationConfig
 from database_services.database_service import DatabaseService
 from openai_services.openai_service import OpenAIService
+from schema.schema import (User, UserUpdate, UserUpdatePass, Password,
+                           Completions, ChatCompletions,
+                           Embeddings, FineTunes, ADMIN_LIST)
 
 from authentication_services.authentication_service import AuthenticationService
-from utils.http_exceptions import limit_exception, forbidden_exception
+from utils.http_exceptions import (limit_exception, forbidden_exception,
+                                   owner_removal_exception,
+                                   self_removal_exception,
+                                   admin_remove_admin_exception,
+                                   admin_add_owner_exception,
+                                   admin_add_admin_exception,
+                                   admin_update_owner_exception,
+                                   admin_update_admin_exception,
+                                   wrong_password_exception)
 
 
 service_config = ServiceConfig()
@@ -22,24 +32,32 @@ database = DatabaseService(database_config=database_config)
 openai_config = OpenAIConfig()
 openai_service = OpenAIService(openai_config=openai_config)
 
-auth_service = AuthenticationService(database_service=database)
+auth_config = AuthenticationConfig()
+auth_service = AuthenticationService(database_service=database,
+                                     auth_config=auth_config)
 app = FastAPI(docs_url=service_config.url_swagger,
               redoc_url=service_config.url_redoc)
 
-@app.post("/admin/token")
+
+@app.post("/admin/token", include_in_schema=False)
 async def login(form_data: OAuth2PasswordRequestForm=Depends()):
     """Login endpoint"""
     result = await database.verify_admin(form_data.username, form_data.password)
     if result['acknowledged'] is False:
-        raise HTTPException(status_code=401,
-                            detail="Incorrect username or password")
+        raise wrong_password_exception
+
+    result.pop("_id")
+    result.pop("password")
+    result.pop("api_key")
+    result.pop("acknowledged")
+
     return {
-        "access_token": auth_service.generate_access_token(result['username']),
-        "refresh_token": auth_service.generate_refresh_token(result['username'])
+        "access_token": auth_service.generate_access_token(result),
+        "refresh_token": auth_service.generate_refresh_token(result)
     }
 
-@app.post("/admin/add", dependencies=[Depends(auth_service.validate_token)])
-async def add(user: User):
+@app.post("/admin/users", include_in_schema=False)
+async def add(user: User,  user_info: str=Depends(auth_service.validate_token)):
     """Add new user to database
     
     Args:
@@ -49,6 +67,15 @@ async def add(user: User):
         dict: result of adding data to database.
     
     """
+    if user_info['sub']['privilege'] not in ADMIN_LIST:
+        raise forbidden_exception
+
+    if user_info['sub']['privilege'] == "admin":
+        if user.privilege == "owner":
+            raise admin_add_owner_exception
+        if user.privilege == "admin":
+            raise admin_add_admin_exception
+
     result = await database.add_new_user(user)
     if result['acknowledged'] is False:
         raise HTTPException(status_code=result['status_code'],
@@ -57,8 +84,9 @@ async def add(user: User):
         return HTTPException(status_code=result['status_code'],
                              detail=result['API_key'])
 
-@app.put("/admin/update", dependencies=[Depends(auth_service.validate_token)])
-async def update(user: UserUpdate):
+@app.put("/admin/users", include_in_schema=False)
+async def update(user: UserUpdate,
+                 user_info: str=Depends(auth_service.validate_token)):
     """Update user in database
     
     Args:
@@ -68,6 +96,17 @@ async def update(user: UserUpdate):
         dict: result of updating data in database.
     
     """
+    if user_info['sub']['privilege'] not in ADMIN_LIST:
+        raise forbidden_exception
+
+    result = await database.retrieve_user(user.user_id, 'user_id')
+    if result['acknowledged']:
+        if user_info['sub']['privilege'] == "admin":
+            if result['data']['privilege'] == "owner":
+                raise admin_update_owner_exception
+            if result['data']['privilege'] == "admin":
+                raise admin_update_admin_exception
+
     result = await database.edit_user(user)
     if result['acknowledged'] is False:
         raise HTTPException(status_code=result['status_code'],
@@ -75,44 +114,9 @@ async def update(user: UserUpdate):
     else:
         return await database.retrieve_user(user.user_id, 'user_id')
 
-@app.get("/admin/get/{user_id}",
-         dependencies=[Depends(auth_service.validate_token)])
-async def get_user(user_id: str):
-    """Get user in database
-
-    Args:
-        user_id (str): ID of the user data.
-
-    Returns:
-        dict: result of user data in database.
-
-    """
-    result = await database.retrieve_user(user_id, 'user_id')
-    if result['acknowledged'] is False:
-        raise HTTPException(status_code=result['status_code'],
-                            detail=result['message'])
-    else:
-        return result['data']
-
-@app.get("/admin/get",
-         dependencies=[Depends(auth_service.validate_token)])
-async def get_all_user():
-    """Get all users in database
-
-    Args:
-        None
-
-    Returns:
-        dict: result of user data in database.
-
-    """
-    result = await database.find_all_users()
-
-    return result
-
-@app.delete("/admin/delete/{user_id}", 
-            dependencies=[Depends(auth_service.validate_token)])
-async def delete_user(user_id: str):
+@app.delete("/admin/users/{user_id}", include_in_schema=False)
+async def delete_user(user_id: str,
+                      user_info: str=Depends(auth_service.validate_token)):
     """Delete user in database
 
     Args:
@@ -122,6 +126,20 @@ async def delete_user(user_id: str):
         dict: result of user data in database.
 
     """
+    if user_info['sub']['privilege'] not in ADMIN_LIST:
+        raise forbidden_exception
+
+    if user_info['sub']['user_id'] == user_id:
+        raise self_removal_exception
+
+    result = await database.retrieve_user(user_id, 'user_id')
+    if result['acknowledged']:
+        if result['data']['privilege'] == "owner":
+            raise owner_removal_exception
+        if user_info['sub']['privilege'] == "admin" and \
+            result['data']['privilege'] == "admin":
+            raise admin_remove_admin_exception
+
     result = await database.delete_user(user_id, 'user_id')
     if result['acknowledged'] is False:
         raise HTTPException(status_code=result['status_code'],
@@ -130,10 +148,91 @@ async def delete_user(user_id: str):
         await database.delete_request_ts_record(user_id)
         return result['data']
 
-@app.get("/admin/get_record/{user_id}",
-         dependencies=[Depends(auth_service.validate_token)])
+@app.put("/admin/users/me/password", include_in_schema=False)
+async def change_password(password: Password,
+                          user_info: str=Depends(auth_service.validate_token)):
+    """Update user password
+
+    Args:
+        user_id (str): ID of the user data.
+
+    Returns:
+        dict: result of user data in database.
+
+    """
+
+    result = await database.verify_admin(username=user_info['sub']['user_id'],
+                                         password=password.old_password)
+    if result['acknowledged'] is False:
+        raise wrong_password_exception
+
+    user = UserUpdatePass(user_id=user_info['sub']['user_id'],
+                          password=password.password)
+
+    result = await database.edit_user(user)
+    if result['acknowledged'] is False:
+        raise HTTPException(status_code=result['status_code'],
+                            detail=result['message'])
+    else:
+        return await database.retrieve_user(user.user_id, 'user_id')
+
+@app.get("/admin/users", include_in_schema=False)
+async def get_all_user(user_info: str=Depends(auth_service.validate_token)):
+    """Get all users in database
+
+    Args:
+        None
+
+    Returns:
+        dict: result of user data in database.
+
+    """
+    if user_info['sub']['privilege'] not in ADMIN_LIST:
+        raise forbidden_exception
+    result = await database.find_all_users()
+
+    return result
+
+@app.get("/admin/users/me", include_in_schema=False)
+async def get_current_user(user_info: str=Depends(auth_service.validate_token)):
+    """Get the current user in database
+
+    Args:
+        None
+
+    Returns:
+        dict: result of user data in database.
+
+    """
+    return user_info['sub']
+
+@app.get("/admin/users/{user_id}", include_in_schema=False)
+async def get_user(user_id: str,
+                   user_info: str=Depends(auth_service.validate_token)):
+    """Get user in database
+
+    Args:
+        user_id (str): ID of the user data.
+
+    Returns:
+        dict: result of user data in database.
+
+    """
+    if user_info['sub']['privilege'] not in ADMIN_LIST and \
+        user_info['sub']['user_id'] != user_id:
+        raise forbidden_exception
+
+    result = await database.retrieve_user(user_id, 'user_id')
+    if result['acknowledged'] is False:
+        raise HTTPException(status_code=result['status_code'],
+                            detail=result['message'])
+    else:
+        return result['data']
+
+@app.get("/admin/users/{user_id}/records", include_in_schema=False)
 async def get_record(user_id: str, endpoint: str, day_from: float,
-                     day_to: float=0.0, slice: str="hour"):
+                     day_to: float=0.0, slice: str="hour",
+                     user_info: str=Depends(auth_service.validate_token)):
     """Get ts record in database
 
     Args:
@@ -147,10 +246,55 @@ async def get_record(user_id: str, endpoint: str, day_from: float,
         list: result of ts data in database.
 
     """
+    if user_info['sub']['privilege'] not in ADMIN_LIST and \
+        user_info['sub']['user_id'] != user_id:
+        raise forbidden_exception
+
     result = await database.get_ts_dates(user_id, endpoint=endpoint,
                                          day_from=day_from, day_to=day_to,
                                          slice=slice)
+
+    if result['acknowledged'] is False:
+        raise HTTPException(status_code=result['status_code'],
+                            detail=result['message'])
+
+    return result["data"]
+
+@app.get("/admin/init", dependencies=[Depends(auth_service.init_key_auth)],
+         include_in_schema=False)
+async def is_initialized():
+    """Check if initialized
+
+    Args:
+        None
+
+    Returns:
+        bool: if database is intialized.
+
+    """
+    result = await database.is_init()
+
     return result
+
+@app.post("/admin/init", dependencies=[Depends(auth_service.init_key_auth)],
+          include_in_schema=False)
+async def initialize(user: User):
+    """Initialize database
+
+    Args:
+        user (User): Input owner data.
+
+    Returns:
+        dict: result of owner data in database.
+
+    """
+    result = await database.init(user=user)
+    if result['acknowledged'] is False:
+        raise HTTPException(status_code=result['status_code'],
+                            detail=result['message'])
+    else:
+        return HTTPException(status_code=result['status_code'],
+                             detail=result['API_key'])
 
 @app.get("/get")
 async def retrieve_user(user_info: str=Depends(auth_service.api_key_auth)):
@@ -236,7 +380,7 @@ async def embeddings(embeddings_args: Embeddings,
     """
     if user_info['request_limit'] == 0:
         raise limit_exception
-    
+
     if not user_info['permissions']['embeddings']:
         raise forbidden_exception
 
@@ -393,4 +537,4 @@ async def list_fine_tunes(user_info: str=Depends(auth_service.api_key_auth)):
 
 
 if __name__ == "__main__":
-   uvicorn.run("app:app", host=service_config.host, port=service_config.port)
+    uvicorn.run("app:app", host=service_config.host, port=service_config.port)
